requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cssselect2==0.3.0
Cython==0.29.16
defusedxml==0.6.0
Django==2.2.10
django-ckeditor==5.9.0
django-datatables-view==1.19.1
django-filter==2.2.0
django-js-asset==1.2.2
django-tables2==2.3.1
django-tables2-bootstrap4==2.0.0b5
djangorestframework==3.11.0
docopt==0.6.2
docx==0.2.4
docxtpl==0.9.0
et-xmlfile==1.0.1
ffmpeg-python==0.2.0
future==0.18.2
googletrans==2.4.0
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.14.1
jsonfield==3.1.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
nltk==3.5
num2words==0.5.10
numpy==1.18.2
openpyxl==3.0.3
packaging==20.3
pathlib==1.0.1
pdf2image==1.12.1
Pillow==7.1.1
preview-generator==0.12
psycopg2-binary==2.8.5
pycparser==2.20
pyexif==0.5.0
pyexifinfo==0.4.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-docx==0.8.10
python-magic==0.4.15
pytz==2019.3
regex==2020.4.4
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
textblob==0.15.3
tinycss2==1.0.2
tqdm==4.45.0
Unidecode==1.1.1
urllib3==1.25.8
Wand==0.5.9
webencodings==0.5.1
whitenoise==5.0.1
XlsxWriter==1.2.8
xvfbwrapper==0.2.9
